[IA-3348] rename runtime-common to runtime-common-components; make some changes to AzureComputeModal tests
