+++ src/pages/workspaces/workspace/analysis/RuntimeManager.js
@@ -16,7 +16,7 @@ import { clearNotification, notify } from 'src/libs/notifications'
 import { useOnMount, usePrevious } from 'src/libs/react-utils'
 import { errorNotifiedApps, errorNotifiedRuntimes } from 'src/libs/state'
 import * as Utils from 'src/libs/utils'
-import { appLauncherTabName, GalaxyLaunchButton, GalaxyWarning } from 'src/pages/workspaces/workspace/analysis/runtime-common'
+import { appLauncherTabName, GalaxyLaunchButton, GalaxyWarning } from 'src/pages/workspaces/workspace/analysis/runtime-common-components'
 import { cloudProviders, getCurrentApp, getCurrentRuntime } from 'src/pages/workspaces/workspace/analysis/runtime-utils'
 import { appTools, toolLabels } from 'src/pages/workspaces/workspace/analysis/tool-utils'
