+++ src/pages/workspaces/workspace/analysis/modals/AzureComputeModal.js
@@ -19,7 +19,7 @@ import { useOnMount } from 'src/libs/react-utils'
 import * as Utils from 'src/libs/utils'
 import { getAzureComputeCostEstimate, getAzureDiskCostEstimate } from 'src/pages/workspaces/workspace/analysis/cost-utils'
 import { WarningTitle } from 'src/pages/workspaces/workspace/analysis/modals/WarningTitle'
-import { computeStyles } from 'src/pages/workspaces/workspace/analysis/runtime-common'
+import { computeStyles } from 'src/pages/workspaces/workspace/analysis/runtime-common-components'
 import {
   getCurrentRuntime, getIsRuntimeBusy
 } from 'src/pages/workspaces/workspace/analysis/runtime-utils'
