+++ src/pages/workspaces/workspace/analysis/modals/GalaxyModal.js
@@ -17,7 +17,7 @@ import * as Style from 'src/libs/style'
 import * as Utils from 'src/libs/utils'
 import { getGalaxyComputeCost, getGalaxyDiskCost } from 'src/pages/workspaces/workspace/analysis/cost-utils'
 import { WarningTitle } from 'src/pages/workspaces/workspace/analysis/modals/WarningTitle'
-import { computeStyles, GalaxyLaunchButton, GalaxyWarning, RadioBlock, SaveFilesHelpGalaxy } from 'src/pages/workspaces/workspace/analysis/runtime-common'
+import { computeStyles, GalaxyLaunchButton, GalaxyWarning, RadioBlock, SaveFilesHelpGalaxy } from 'src/pages/workspaces/workspace/analysis/runtime-common-components'
 import {
   findMachineType, getCurrentApp, getCurrentAppDataDisk, getCurrentAttachedDataDisk, pdTypes
 } from 'src/pages/workspaces/workspace/analysis/runtime-utils'
