+++ src/pages/workspaces/workspace/analysis/runtime-common-components.js
